WarCountyBox moved to package "box"
