+++ src/main/resources/box/WarCountryBox.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.layout.GridPane?>
 
-<GridPane prefHeight="651.0" prefWidth="300.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.40" fx:controller="ee.tkasekamp.vickywaranalyzer.controller.WarCountryBox">
+<GridPane prefHeight="651.0" prefWidth="300.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.40" fx:controller="ee.tkasekamp.vickywaranalyzer.controller.box.WarCountryBox">
 	<children>
 		<HBox id="HBox" alignment="CENTER" spacing="5.0" GridPane.columnIndex="0" GridPane.rowIndex="0">
 			<children>
